chore: more cleanup of tests
